+++ lib/hkscs1999.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 1999-2006, 2012 Free Software Foundation, Inc.
+ * Copyright (C) 1999-2006, 2012, 2016 Free Software Foundation, Inc.
  * This file is part of the GNU LIBICONV Library.
  *
  * The GNU LIBICONV Library is free software; you can redistribute it
@@ -849,7 +849,7 @@ static const ucs4_t hkscs1999_2uni_upages[973] = {
 };
 
 static int
-hkscs1999_mbtowc (conv_t conv, ucs4_t *pwc, const unsigned char *s, int n)
+hkscs1999_mbtowc (conv_t conv, ucs4_t *pwc, const unsigned char *s, size_t n)
 {
   unsigned char c1 = s[0];
   if ((c1 >= 0x88 && c1 <= 0x8b) || (c1 >= 0x8d && c1 <= 0xa0) || (c1 >= 0xc6 && c1 <= 0xc8) || (c1 >= 0xf9 && c1 <= 0xfe)) {
@@ -2949,7 +2949,7 @@ static const Summary16 hkscs1999_uni2indx_page2f8[30] = {
 };
 
 static int
-hkscs1999_wctomb (conv_t conv, unsigned char *r, ucs4_t wc, int n)
+hkscs1999_wctomb (conv_t conv, unsigned char *r, ucs4_t wc, size_t n)
 {
   if (n >= 2) {
     const Summary16 *summary = NULL;
