+++ lib/jisx0212.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 1999-2001, 2012 Free Software Foundation, Inc.
+ * Copyright (C) 1999-2001, 2012, 2016 Free Software Foundation, Inc.
  * This file is part of the GNU LIBICONV Library.
  *
  * The GNU LIBICONV Library is free software; you can redistribute it
@@ -910,7 +910,7 @@ static const unsigned short jisx0212_2uni_page30[5801] = {
 };
 
 static int
-jisx0212_mbtowc (conv_t conv, ucs4_t *pwc, const unsigned char *s, int n)
+jisx0212_mbtowc (conv_t conv, ucs4_t *pwc, const unsigned char *s, size_t n)
 {
   unsigned char c1 = s[0];
   if ((c1 == 0x22) || (c1 >= 0x26 && c1 <= 0x27) || (c1 >= 0x29 && c1 <= 0x2b) || (c1 >= 0x30 && c1 <= 0x6d)) {
@@ -2153,7 +2153,7 @@ static const Summary16 jisx0212_uni2indx_pageff[6] = {
 };
 
 static int
-jisx0212_wctomb (conv_t conv, unsigned char *r, ucs4_t wc, int n)
+jisx0212_wctomb (conv_t conv, unsigned char *r, ucs4_t wc, size_t n)
 {
   if (n >= 2) {
     const Summary16 *summary = NULL;
